Add data visualization

+++ data.cpp
@@ -1,10 +1,10 @@
 #include "data.h"
 #include <cmath>
 
-int City::getX() const{
+double City::getX() const{
 	return x;
 }
-int City::getY() const{
+double City::getY() const{
 	return y;
 }
 double City::computeDistance(const City& city) const{
