+++ data.h
@@ -3,17 +3,17 @@
 #include <iostream>
 class City{
 	private:
-		int x;
-		int y;
+		double x;
+		double y;
 	public:
-		City(int x, int y):x(x),y(y){};
+		City(double x, double y):x(x),y(y){};
 		City(){
 			x = -1;
 			y = -1;
 		};
 		double computeDistance(const City&) const;
-		int getX() const;
-		int getY() const;
+		double getX() const;
+		double getY() const;
 };
 static bool operator==(const City& lhs, const City& rhs){
 	return lhs.getX() == rhs.getX() && lhs.getY() == rhs.getY();
